@@ -9,93 +9,71 @@ class PusherMessage
 {
     /**
      * The device platform (iOS/Android).
-     *
-     * @var string
      */
-    protected $platform = 'iOS';
+    protected string $platform = 'iOS';
 
     /**
      * The message title.
-     *
-     * @var string
      */
-    protected $title;
-
-    /**
-     * The message body.
-     *
-     * @var string
-     */
-    protected $body;
+    protected string|null $title = null;
 
     /**
      * The phone number the message should be sent from.
-     *
-     * @var string
      */
-    protected $sound = 'default';
+    protected string $sound = 'default';
 
     /**
      * The message icon (Android).
-     *
-     * @var string
      */
-    protected $icon;
+    protected string|null $icon = null;
 
     /**
      * The number to display next to the push notification (iOS).
-     *
-     * @var int
      */
-    protected $badge;
+    protected int|null $badge = null;
 
     /**
      * URL to follow on notification click.
      */
-    protected $link;
+    protected string|null $link = null;
 
     /**
      * Extra options that will get added to the message.
-     *
-     * @var array
      */
-    protected $options = [];
+    protected array $options = [];
 
     /**
      * An extra message to the other platform.
      *
      * @var
      */
-    protected $extraMessage;
+    protected PusherMessage|null $extraMessage = null;
 
     /**
-     * @param string $body
-     *
+     * @param  string  $body
+     */
+    public function __construct(protected string $body = '')
+    {
+    }
+
+    /**
+     * @param  string  $body
      * @return static
      */
-    public static function create($body = '')
+    public static function create(string $body = ''): static
     {
         return new static($body);
     }
 
     /**
-     * @param string $body
-     */
-    public function __construct($body = '')
-    {
-        $this->body = $body;
-    }
-
-    /**
      * Set the platform [iOS/Android].
      *
-     * @param string $platform
-     *
+     * @param  string  $platform
      * @return $this
      *
      * @throws \NotificationChannels\PusherPushNotifications\Exceptions\CouldNotCreateMessage
      */
-    public function platform($platform)
+    public function platform(string $platform): self
     {
         if (! in_array($platform, ['iOS', 'Android', 'web'])) {
             throw CouldNotCreateMessage::invalidPlatformGiven($platform);
@@ -111,7 +89,7 @@ class PusherMessage
      *
      * @return $this
      */
-    public function iOS()
+    public function iOS(): self
     {
         $this->platform = 'iOS';
 
@@ -123,7 +101,7 @@ class PusherMessage
      *
      * @return $this
      */
-    public function android()
+    public function android(): self
     {
         $this->platform = 'Android';
 
@@ -135,7 +113,7 @@ class PusherMessage
      *
      * @return $this
      */
-    public function web()
+    public function web(): self
     {
         $this->platform = 'web';
 
@@ -145,10 +123,12 @@ class PusherMessage
     /**
      * Set an extra message to be sent to Android.
      *
-     * @param \NotificationChannels\PusherPushNotifications\PusherMessage $message
+     * @param  \NotificationChannels\PusherPushNotifications\PusherMessage  $message
      * @return $this
+     *
+     * @throws CouldNotCreateMessage
      */
-    public function withAndroid(self $message)
+    public function withAndroid(self $message): self
     {
         $this->withExtra($message->android());
 
@@ -158,10 +138,12 @@ class PusherMessage
     /**
      * Set an extra message to be sent to iOS.
      *
-     * @param \NotificationChannels\PusherPushNotifications\PusherMessage $message
+     * @param  \NotificationChannels\PusherPushNotifications\PusherMessage  $message
      * @return $this
+     *
+     * @throws CouldNotCreateMessage
      */
-    public function withiOS(self $message)
+    public function withiOS(self $message): self
     {
         $this->withExtra($message->iOS());
 
@@ -171,10 +153,12 @@ class PusherMessage
     /**
      * Set an extra message to be sent to web.
      *
-     * @param \NotificationChannels\PusherPushNotifications\PusherMessage $message
+     * @param  \NotificationChannels\PusherPushNotifications\PusherMessage  $message
      * @return $this
+     *
+     * @throws CouldNotCreateMessage
      */
-    public function withWeb(self $message)
+    public function withWeb(self $message): self
     {
         $this->withExtra($message->web());
 
@@ -184,12 +168,14 @@ class PusherMessage
     /**
      * Set an extra message to be sent to another platform.
      *
-     * @param \NotificationChannels\PusherPushNotifications\PusherMessage $message
+     * @param  \NotificationChannels\PusherPushNotifications\PusherMessage  $message
      * @return void
+     *
+     * @throws CouldNotCreateMessage
      */
-    private function withExtra(self $message)
+    private function withExtra(self $message): void
     {
-        if ($message->getPlatform() == $this->platform) {
+        if ($message->getPlatform() === $this->platform) {
             throw CouldNotCreateMessage::platformConflict($this->platform);
         }
 
@@ -199,11 +185,10 @@ class PusherMessage
     /**
      * Set the message title.
      *
-     * @param string $value
-     *
+     * @param  string  $value
      * @return $this
      */
-    public function title($value)
+    public function title(string $value): self
     {
         $this->title = $value;
 
@@ -213,11 +198,10 @@ class PusherMessage
     /**
      * Set the message body.
      *
-     * @param string $value
-     *
+     * @param  string  $value
      * @return $this
      */
-    public function body($value)
+    public function body(string $value): self
     {
         $this->body = $value;
 
@@ -227,11 +211,10 @@ class PusherMessage
     /**
      * Set the message sound (Android).
      *
-     * @param string $value
-     *
+     * @param  string  $value
      * @return $this
      */
-    public function sound($value)
+    public function sound(string $value): self
     {
         $this->sound = $value;
 
@@ -241,11 +224,10 @@ class PusherMessage
     /**
      * Set the message icon (Android).
      *
-     * @param string $value
-     *
+     * @param  string  $value
      * @return $this
      */
-    public function icon($value)
+    public function icon(string $value): self
     {
         $this->icon = $value;
 
@@ -255,13 +237,12 @@ class PusherMessage
     /**
      * Set the message badge (iOS).
      *
-     * @param int $value
-     *
+     * @param  int  $value
      * @return $this
      */
-    public function badge($value)
+    public function badge(int $value): self
     {
-        $this->badge = (int) $value;
+        $this->badge = $value;
 
         return $this;
     }
@@ -269,11 +250,10 @@ class PusherMessage
     /**
      * Set the message link.
      *
-     * @param string $value
-     *
+     * @param  string  $value
      * @return $this
      */
-    public function link($value)
+    public function link(string $value): self
     {
         $this->link = $value;
 
@@ -281,12 +261,11 @@ class PusherMessage
     }
 
     /**
-     * @param string $key
-     * @param mixed $value
-     *
+     * @param  string  $key
+     * @param  mixed  $value
      * @return $this
      */
-    public function setOption($key, $value)
+    public function setOption(string $key, mixed $value): self
     {
         $this->options[$key] = $value;
 
@@ -298,16 +277,13 @@ class PusherMessage
      *
      * @return array
      */
-    public function toArray()
+    public function toArray(): array
     {
-        switch ($this->platform) {
-            case 'Android':
-                return $this->toAndroid();
-            case 'web':
-                return $this->toWeb();
-            default:
-                return $this->toiOS();
-        }
+        return match ($this->platform) {
+            'Android' => $this->toAndroid(),
+            'web'     => $this->toWeb(),
+            default   => $this->toiOS(),
+        };
     }
 
     /**
@@ -315,7 +291,7 @@ class PusherMessage
      *
      * @return array
      */
-    public function toiOS()
+    public function toiOS(): array
     {
         $message = [
             'apns' => [
@@ -340,7 +316,7 @@ class PusherMessage
      *
      * @return array
      */
-    public function toAndroid()
+    public function toAndroid(): array
     {
         $message = [
             'fcm' => [
@@ -363,7 +339,7 @@ class PusherMessage
      *
      * @return array
      */
-    public function toWeb()
+    public function toWeb(): array
     {
         $message = [
             'web' => [
@@ -387,7 +363,7 @@ class PusherMessage
      *
      * @return string
      */
-    public function getPlatform()
+    public function getPlatform(): string
     {
         return $this->platform;
     }
@@ -397,7 +373,7 @@ class PusherMessage
      *
      * @param $message
      */
-    private function formatMessage(&$message)
+    private function formatMessage(&$message): void
     {
         if ($this->extraMessage) {
             $message = array_merge($message, $this->extraMessage->toArray());
