@@ -7,19 +7,19 @@ use Exception;
 class CouldNotCreateMessage extends Exception
 {
     /**
-     * @param string $platform
+     * @param  string  $platform
      * @return static
      */
-    public static function invalidPlatformGiven($platform)
+    public static function invalidPlatformGiven(string $platform): static
     {
         return new static("Platform `{$platform}` is invalid. It should be either `iOS` or `Android`.");
     }
 
     /**
-     * @param $platform
+     * @param  string  $platform
      * @return static
      */
-    public static function platformConflict($platform)
+    public static function platformConflict(string $platform): static
     {
         return new static("You are trying to send an extra message to `{$platform}` while the original message is to `{$platform}`.");
     }
